+++ FYDPSleepyZzz/RollingAverage.h
@@ -4,13 +4,13 @@
 class RollingAverage
 {
   public:
-    RollingAverage(int numReadings);
+    RollingAverage(byte numReadings);
     ~RollingAverage();
     void add_value(float new_value, float* new_average);
   private:
-    int   _numReadings;
+    byte   _numReadings;
     float*  readings;                  // the readings from the analog input
-    int   readIndex = 0;              // the index of the current reading
+    byte   readIndex = 0;              // the index of the current reading
     float total = 0;                  // the running total
     float average = 0;                // the average
 };
