+++ FYDPSleepyZzz/TMP36.ino
@@ -1,6 +1,6 @@
 void get_temp_in_celsius(float* celsius_temp)
 {
-    int rawTempVoltage = analogRead(tempPin); 
+    short int rawTempVoltage = analogRead(tempPin); 
 //    float tempMilliVolts = (rawTempVoltage * ((5000/1024)/1000));
     float tempMilliVolts = (rawTempVoltage * 0.0048828125);
     *celsius_temp = (tempMilliVolts - 0.5) * 100;
